+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## `1.0.2`
 
   - 新增 `package.json` 中配置 `ant-design` 主题颜色
+  - 新增模块热替换插件 `react-hot-loader`
   - 针对 `webpack.config.js` 中潜在问题做了优化和修复
   - 删除多线程压缩插件 `webpack-parallel-uglify-plugin`, 由于可能会在压缩后 `manifest.js` 的 `hash` 值不一致，导致页面访问报错
   - 多线程打包功能并不理想，删除 `happypack` ，原有的多线程 `happypack` 方案移动到新分支 `webpack-antd-happypack`